requirements.txt
datasets==3.3.2
faiss-gpu==1.7.2
huggingface-hub==0.29.1
pandas==2.2.3
numpy==1.24.0
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
streamlit==1.42.2
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0